Move non-user scripts into scripts directory
