+++ .github/workflows/ci-build.yml
@@ -175,7 +175,7 @@ jobs:
 
       - name: Merge flash image
         run: |
-          python merge_image.py ${{ matrix.board }} &&
+          python scripts/merge_image.py ${{ matrix.board }} &&
             mv merged.bin OpenShock.${{ matrix.board }}.bin
 
       - name: Upload merged flashable image
