+++ platformio.ini
@@ -24,8 +24,8 @@ lib_deps =
 board_build.partitions = huge_app.csv # Overridden per board
 board_build.filesystem = littlefs
 extra_scripts =
-  pre:embed_env_vars.py
-  post:build_frontend.py
+  pre:scripts/embed_env_vars.py
+  post:scripts/build_frontend.py
 
 ; Serial Monitor options
 upload_speed = 921600
