+++ scripts/merge_image.py
